requirements.txt
gevent==1.0.1
greenlet==0.4.5
ipython==2.4.1
pyzmq==14.5.0
